Move otelcolclient package under pkg
