+++ src/cmd/forwarder-agent/app/forwarder_agent.go
@@ -18,12 +18,12 @@ import (
 	gendiodes "code.cloudfoundry.org/go-diodes"
 	"code.cloudfoundry.org/go-loggregator/v9"
 	"code.cloudfoundry.org/go-loggregator/v9/rpc/loggregator_v2"
-	"code.cloudfoundry.org/loggregator-agent-release/src/cmd/forwarder-agent/app/otelcolclient"
 	"code.cloudfoundry.org/loggregator-agent-release/src/pkg/diodes"
 	"code.cloudfoundry.org/loggregator-agent-release/src/pkg/egress"
 	"code.cloudfoundry.org/loggregator-agent-release/src/pkg/egress/syslog"
 	egress_v2 "code.cloudfoundry.org/loggregator-agent-release/src/pkg/egress/v2"
 	v2 "code.cloudfoundry.org/loggregator-agent-release/src/pkg/ingress/v2"
+	"code.cloudfoundry.org/loggregator-agent-release/src/pkg/otelcolclient"
 	"code.cloudfoundry.org/loggregator-agent-release/src/pkg/plumbing"
 	"code.cloudfoundry.org/loggregator-agent-release/src/pkg/timeoutwaitgroup"
 	"google.golang.org/grpc"
