+++ src/pkg/otelcolclient/metric_batcher.go
